+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGE LOG
 
+### v0.3.0
+
+- Breaking: remove reflection functions. No need to use them.
+- Fix: now the package can be used within other modules and can be precompiled within other modules.
+
+### v0.2.1
+
+- Feat: no need to manually call `init`. This function is removed.
+
 ### v0.2.0
 
 - Breaking: new method `init(incremental::Bool=true)` and `gen_scoped_stream_methods(incremental::Bool=true)`.